Add option to retrieve logs once without watching / streaming (#14)
This adds an alternative way to get logs other than the current follow logs option.

It uses the same path to query the API but without the follow=true option, hence it doesn't
require any retry semantics. As the returned logs are plane text, a number of functions 
have been duplicated to handle this.