Fix bug when trying to fetch logs without follow flag (#16)
Sending follow=true in the query of any request to the Kubernetes API changes 
the behaviour of the response significantly. So it's critical that it is not sent if the
request is not intended to stream the result. This PR fixes that so we can easily
distinguish between a dump of the logs and a stream of the logs.