Ignore response payload when delete-call doesn't return a meta.v1.Status
Deleting a core.v1.Service returns the deleted resource instead of the
specified meta.v1.Status. Thus, we ignore the returned payload and
return a dummy meta.v1.Status with the success status code in this case.

This should not change the behaviour from the perspective of the caller
code, since the delete API doesn't habe a return type anyways.

Fixes #27